requirements.txt
# --- Core Application ---
pydantic-ai==1.23.0
langgraph==1.0.4
openai==1.60.1
pydantic==2.12.5
pydantic-settings==2.7.1

# --- Server & Networking ---
fastapi==0.115.2
uvicorn==0.32.0
httpx==0.28.1
aiogram==3.15.0

# --- Database & External Services ---
supabase==2.9.1
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.1.0

# --- Background Tasks ---
celery>=5.4.0
redis>=5.0.0
kombu>=5.3.0
flower>=2.0.0

# --- Monitoring ---
sentry-sdk[celery,fastapi]>=2.0.0

# --- Utilities ---
python-dotenv>=1.0.0

# --- Testing ---
pytest==9.0.1
pytest-asyncio==0.28.0
